// TPT/manage-tradingpt/app/api/users.ts
export interface FreeCustomer {
  customerId: number;
  name: string;
  phoneNumber: string;
  uid: string | null;
  primaryInvestmentType: string | null;
  token: number;
  createdAt: string;
}